Add elasticloadbalance:AddTags permission to deployer role
AWS says

We are reaching out to you regarding an upcoming change that affects your usage
of tags with Elastic Load Balancing (ELB) API calls. On June 1, 2023, we will
be adding an additional layer of security to ELB ‘Create*' API calls where API
callers must have explicit access to add tags in their Identity and Access
Management (IAM) policy [1]. Currently, access to attach tags was implicitly
granted with access to 'Create*' APIs.

We detected that you have used IAM policies that are affected by this change.
To prevent issues, we recommend you update your IAM policy configuration to
allow access to 'elasticloadbalancing:AddTags' when using 'Create*' API calls.
We will be allowing 'Create*' API calls with the current policy to be accepted
until August 30, 2023. After this date, the 'Create*' API call will fail and
return an error if the the attribute is specified and permission is not granted.